Support --metabase_session_id
Metabase lacks proper support for authenticating from automation, which
means we have to exchange a username and password for a session token
before each invocation of this command.

These count as login attempts, which are rate limited. Depending on your
needs, you may hit the rate limit and find your schemas can no longer be
synced.

Providing a --metabase_session_id allows the user to reuse and existing
Metabase session, should they have it, allowing them to avoid the login
rate limits (and in some cases, an annoying email!).