Support setting metabase.number_style on columns
I have no idea why they have made a nested settings field in the column properties, but the behavior is that if we provide a settings entry, it will overwrite all the settings that appear in there in their responses (which are undocumented). So I've broken with the pattern a bit and am copying in the whole settings we were given, to avoid removing any other formatting changes made in the UI and make it easy to add more settings in the future.