Listing dbt tables (models, sources) using `dbt ls`.
The main motivation for doing so is being able to safely get dbt models that are not defined in the property YAML files.
An alternative would've been to parse `models/*.sql` files but that's difficult due to the `alias` config that can override the table name from the default file name.
Using `dbt ls` allows us to safely get the table name.